sam-app/bunq2ynab/requirements.txt
bunq-sdk
requests~=2.23.0
setuptools~=46.1.3
aws-cdk.core
aws-cdk.aws_iam
aws-cdk.aws_sqs
aws-cdk.aws_sns
aws-cdk.aws_sns_subscriptions
aws-cdk.aws_s3
pyopenssl
network
boto3~=1.13.1